refactor(goal): split boundary helpers for lint ratchet
Problem: the Goal boundary implementation grew several grandfathered files beyond the repository's repolint ratchet.

Root cause: run-boundary state, sidecar compatibility helpers, evidence novelty logic, and their tests were added directly to legacy owner files.

Fix: extract the new logic into focused same-package files, reduce the affected owner functions, and keep the existing repolint baseline unchanged.

Verification: go run ./tools/repolint; go test ./...; go test -race ./internal/agent ./internal/control; go vet ./...; desktop go test ./...; frontend pnpm build.
