+++ internal/control/goal_progress.go
@@ -1,0 +1,72 @@
+package control
+
+import "strings"
+
+// noProgressQuota returns deprecated compatibility metadata for old clients.
+func noProgressQuota(class string) int {
+	switch class {
+	case budgetClassResearch:
+		return 10
+	case budgetClassWrite:
+		return 6
+	default:
+		return defaultNoProgressLimit
+	}
+}
+
+func resolvedNoProgressLimit(persisted int, class string) int {
+	if persisted > 0 && persisted != defaultNoProgressLimit {
+		return persisted
+	}
+	return noProgressQuota(class)
+}
+
+// mergeGoalProgressEvidence updates the bounded Goal-scoped novelty window.
+func mergeGoalProgressEvidence(existing, observed []string) ([]string, bool) {
+	if len(existing) > maxGoalProgressEvidence {
+		existing = existing[len(existing)-maxGoalProgressEvidence:]
+	}
+	if len(observed) > maxGoalProgressEvidence {
+		observed = observed[len(observed)-maxGoalProgressEvidence:]
+	}
+	out := make([]string, 0, min(len(existing)+len(observed), maxGoalProgressEvidence))
+	seen := make(map[string]struct{}, min(len(existing)+len(observed), maxGoalProgressEvidence))
+	appendValid := func(sig string) bool {
+		sig = strings.TrimSpace(sig)
+		if sig == "" || len(sig) > 128 {
+			return false
+		}
+		if _, ok := seen[sig]; ok {
+			return false
+		}
+		seen[sig] = struct{}{}
+		out = append(out, sig)
+		return true
+	}
+	for _, sig := range existing {
+		appendValid(sig)
+	}
+	progressed := false
+	for _, sig := range observed {
+		if appendValid(sig) {
+			progressed = true
+		}
+	}
+	if len(out) > maxGoalProgressEvidence {
+		out = append([]string(nil), out[len(out)-maxGoalProgressEvidence:]...)
+	}
+	return out, progressed
+}
+
+func (g *goalMachine) observeGoalProgress(in goalAdvanceInput) {
+	progressed := false
+	g.progressEvidence, progressed = mergeGoalProgressEvidence(g.progressEvidence, in.progressEvidence)
+	if in.report != nil && in.report.status != GoalStatusRunning {
+		progressed = true
+	}
+	if progressed {
+		g.noProgressTurns = 0
+		return
+	}
+	g.noProgressTurns++
+}
