+++ internal/control/goal_runtime_test.go
@@ -3,7 +3,6 @@ package control
 import (
 	"context"
 	"encoding/json"
-	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
@@ -194,167 +193,6 @@ func TestEvaluatorCompleteStillGatedByReadiness(t *testing.T) {
 // TestTurnTokenNoProgressPausesAndResumeExtendsBudget covers the outer turn
 // budget, observational no-progress state, and the resume extension contract.
 // Token hard limits no longer pause goals.
-func TestNoProgressQuotaByBudgetClass(t *testing.T) {
-	cases := map[string]int{
-		budgetClassResearch: 10,
-		budgetClassWrite:    6,
-		budgetClassSimple:   defaultNoProgressLimit,
-	}
-	for class, want := range cases {
-		if got := noProgressQuota(class); got != want {
-			t.Fatalf("noProgressQuota(%q) = %d, want %d", class, got, want)
-		}
-	}
-	if got := resolvedNoProgressLimit(defaultNoProgressLimit, budgetClassResearch); got != 10 {
-		t.Fatalf("legacy research limit = %d, want 10", got)
-	}
-	if got := resolvedNoProgressLimit(7, budgetClassWrite); got != 7 {
-		t.Fatalf("custom write limit = %d, want 7", got)
-	}
-}
-
-func TestMergeGoalProgressEvidenceIsNovelAndBounded(t *testing.T) {
-	observed := make([]string, maxGoalProgressEvidence+25)
-	for i := range observed {
-		observed[i] = fmt.Sprintf("sig-%03d", i)
-	}
-	got, progressed := mergeGoalProgressEvidence([]string{"old"}, observed)
-	if !progressed || len(got) != maxGoalProgressEvidence {
-		t.Fatalf("merge = progressed:%v len:%d, want true/%d", progressed, len(got), maxGoalProgressEvidence)
-	}
-	if repeat, advanced := mergeGoalProgressEvidence(got, got); advanced || len(repeat) != maxGoalProgressEvidence {
-		t.Fatalf("exact repeat = advanced:%v len:%d", advanced, len(repeat))
-	}
-}
-
-func TestTurnTokenNoProgressPausesAndResumeExtendsBudget(t *testing.T) {
-	newMachine := func() *goalMachine {
-		g := &goalMachine{goal: "fix the parser", status: GoalStatusRunning}
-		g.budgetClass = budgetClassWrite
-		g.turnsLimit = budgetQuota(budgetClassWrite)
-		g.tokensLimit = 0
-		g.noProgressLimit = noProgressQuota(g.budgetClass)
-		return g
-	}
-	in := func(report *goalTurnReport, progress ...string) goalAdvanceInput {
-		return goalAdvanceInput{report: report, progressEvidence: progress}
-	}
-
-	t.Run("turn budget pauses", func(t *testing.T) {
-		g := newMachine()
-		for i := range g.turnsLimit {
-			// Progress changes every turn so only the turn budget can fire.
-			res := g.advance(in(&goalTurnReport{status: GoalStatusRunning, reason: "keep going"}, "sig-"+fmt.Sprint(i)))
-			if res.cont && i == g.turnsLimit-1 {
-				t.Fatal("last turn must pause")
-			}
-		}
-		if g.status != GoalStatusBlocked || g.stopCause != stopCauseBudgetTurns {
-			t.Fatalf("machine = (%q, %q), want blocked+budget_turns", g.status, g.stopCause)
-		}
-	})
-
-	t.Run("token usage never pauses", func(t *testing.T) {
-		g := newMachine()
-		// Even a huge observational total must not stop the goal.
-		g.tokensUsed = 900_000
-		res := g.advance(in(&goalTurnReport{status: GoalStatusRunning, reason: "keep going"}, "new-evidence"))
-		if !res.cont {
-			t.Fatal("token usage must not pause the goal")
-		}
-		if g.stopCause != "" {
-			t.Fatalf("stop cause = %q, want empty", g.stopCause)
-		}
-	})
-
-	t.Run("no-progress is observational", func(t *testing.T) {
-		g := newMachine()
-		for range g.noProgressLimit {
-			if res := g.advance(in(&goalTurnReport{status: GoalStatusRunning, reason: "still working"})); !res.cont {
-				t.Fatalf("legacy 4/6/10 metadata paused Goal: %+v", g)
-			}
-		}
-		if g.status != GoalStatusRunning || g.stopCause != "" || g.noProgressTurns != g.noProgressLimit {
-			t.Fatalf("machine = (%q, %q, %d), want running with observational streak", g.status, g.stopCause, g.noProgressTurns)
-		}
-	})
-
-	t.Run("host-verifiable progress resets no-progress", func(t *testing.T) {
-		g := newMachine()
-		for range g.noProgressLimit - 1 {
-			g.advance(in(&goalTurnReport{status: GoalStatusRunning, reason: "still working"}))
-		}
-		res := g.advance(in(&goalTurnReport{status: GoalStatusRunning, reason: "progress!"}, "new-evidence"))
-		if !res.cont {
-			t.Fatalf("progress must reset the stall counter: %+v", g)
-		}
-		if g.noProgressTurns != 0 {
-			t.Fatalf("noProgressTurns = %d, want 0", g.noProgressTurns)
-		}
-	})
-
-	t.Run("exact repeated evidence does not reset no-progress", func(t *testing.T) {
-		g := newMachine()
-		g.advance(in(&goalTurnReport{status: GoalStatusRunning, reason: "read"}, "same-read"))
-		for range g.noProgressLimit {
-			g.advance(in(&goalTurnReport{status: GoalStatusRunning, reason: "read again"}, "same-read"))
-		}
-		if g.status != GoalStatusRunning || g.stopCause != "" || g.noProgressTurns != g.noProgressLimit {
-			t.Fatalf("repeated evidence should remain observational: status=%q cause=%q streak=%d", g.status, g.stopCause, g.noProgressTurns)
-		}
-	})
-
-	t.Run("research accepts new read evidence but keeps a larger synthesis budget", func(t *testing.T) {
-		g := &goalMachine{
-			goal: "research the protocol", status: GoalStatusRunning,
-			budgetClass: budgetClassResearch, turnsLimit: budgetQuota(budgetClassResearch),
-			noProgressLimit: noProgressQuota(budgetClassResearch),
-		}
-		for i := range 8 {
-			if res := g.advance(in(&goalTurnReport{status: GoalStatusRunning, reason: "researching"}, "read-"+fmt.Sprint(i))); !res.cont {
-				t.Fatalf("new research evidence paused at turn %d", i+1)
-			}
-		}
-		if g.noProgressTurns != 0 || g.noProgressLimit != 10 {
-			t.Fatalf("research runtime = no-progress %d/%d", g.noProgressTurns, g.noProgressLimit)
-		}
-	})
-
-	t.Run("resume extends turn budget once", func(t *testing.T) {
-		g := newMachine()
-		for range g.turnsLimit {
-			g.advance(in(&goalTurnReport{status: GoalStatusRunning, reason: "keep going"}))
-		}
-		beforeLimit := g.turnsLimit
-		_, _, _, resumed, extended := g.resume(nil)
-		if !resumed || !extended {
-			t.Fatalf("resume = (%v, %v), want (true, true)", resumed, extended)
-		}
-		if g.turnsLimit != beforeLimit+budgetQuota(budgetClassWrite) {
-			t.Fatalf("turnsLimit = %d, want %d", g.turnsLimit, beforeLimit+budgetQuota(budgetClassWrite))
-		}
-		if g.tokensLimit != 0 {
-			t.Fatalf("tokensLimit = %d, want 0 (no hard token ceiling)", g.tokensLimit)
-		}
-		if g.budgetExtensions != 1 || g.stopCause != "" || g.noProgressTurns != 0 {
-			t.Fatalf("machine after resume = %+v", g)
-		}
-	})
-
-	t.Run("manual pause resume does not extend", func(t *testing.T) {
-		g := newMachine()
-		g.pauseFor(stopCauseManual, "user paused", nil)
-		beforeLimit := g.turnsLimit
-		_, _, _, resumed, extended := g.resume(nil)
-		if !resumed || extended {
-			t.Fatalf("resume = (%v, %v), want (true, false)", resumed, extended)
-		}
-		if g.turnsLimit != beforeLimit {
-			t.Fatalf("turnsLimit changed on manual resume: %d", g.turnsLimit)
-		}
-	})
-}
-
 // TestGoalTurnRecorderProtocol covers idempotency, upgrades, terminal
 // conflicts, and stale-epoch rejection.
 func TestGoalTurnRecorderProtocol(t *testing.T) {
@@ -464,39 +302,6 @@ func TestGoalTurnRecorderProtocol(t *testing.T) {
 			t.Fatalf("stale usage folded into replacement goal: %d", g.tokensUsed)
 		}
 	})
-}
-
-func TestWireContinueDoesNotKeepExactRepeatedEvidenceAlive(t *testing.T) {
-	g := &goalMachine{
-		goal: "repeat the same read", status: GoalStatusRunning,
-		budgetClass: budgetClassSimple, turnsLimit: budgetQuota(budgetClassSimple),
-		noProgressLimit: noProgressQuota(budgetClassSimple), scopeID: newGoalScopeID(),
-	}
-	advance := func() goalAdvanceResult {
-		epoch := g.continuationEpoch
-		rec := g.newTurnRecorder(g.scopeID, epoch)
-		if _, err := rec.RecordGoalReport(tool.GoalReport{
-			Status: "continue", Reason: "same read", NextAction: "read it again",
-		}); err != nil {
-			t.Fatal(err)
-		}
-		return g.advance(goalAdvanceInput{
-			report: rec.validReport(epoch), progressEvidence: []string{"same-read"}, expectedEpoch: &epoch,
-		})
-	}
-
-	if res := advance(); !res.cont || g.noProgressTurns != 0 {
-		t.Fatalf("first read = %+v runtime=%+v", res, g.runtimeView())
-	}
-	for i := 1; i <= g.noProgressLimit; i++ {
-		res := advance()
-		if i < g.noProgressLimit && !res.cont {
-			t.Fatalf("repeat %d paused early: %+v", i, res)
-		}
-	}
-	if g.status != GoalStatusRunning || g.stopCause != "" || g.noProgressTurns != g.noProgressLimit {
-		t.Fatalf("wire continue should remain active with observational streak: status=%q cause=%q runtime=%+v", g.status, g.stopCause, g.runtimeView())
-	}
 }
 
 // TestGoalUsageTeeAttributesScopedBillableCallsAndExcludesTitle covers the
@@ -634,154 +439,6 @@ func TestGoalLegacyBudgetTokensSidecarAutoResumes(t *testing.T) {
 	}
 	if g2.status != GoalStatusRunning || g2.stopCause != "" {
 		t.Fatalf("second load = %q/%q, want running/empty", g2.status, g2.stopCause)
-	}
-}
-
-func TestGoalLegacyNoProgressSidecarAutoResumesWithoutExtension(t *testing.T) {
-	dir := t.TempDir()
-	path := filepath.Join(dir, "session.jsonl")
-	state := goalState{
-		Goal: "finish the migration", Status: GoalStatusBlocked,
-		StopCause: stopCauseNoProgress, Block: "no host-verifiable progress",
-		BudgetClass: budgetClassWrite, TurnsUsed: 6, TurnsLimit: 20,
-		NoProgressTurns: 6, NoProgressLimit: 6, RequestsUsed: 12,
-	}
-	raw, err := json.Marshal(state)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := os.WriteFile(store.SessionGoalState(path), raw, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	g := &goalMachine{}
-	_, data, migrated, _ := g.restoreFromState(path)
-	if !migrated {
-		t.Fatal("legacy no_progress pause must migrate")
-	}
-	if g.status != GoalStatusRunning || g.stopCause != "" || g.block != "" {
-		t.Fatalf("restored status = %q cause=%q block=%q", g.status, g.stopCause, g.block)
-	}
-	if g.turnsLimit != 20 || g.budgetExtensions != 0 || g.requestsUsed != 12 || g.noProgressTurns != 6 {
-		t.Fatalf("migration changed runtime counters: %+v", g.runtimeView())
-	}
-	var normalized goalState
-	if err := json.Unmarshal(data, &normalized); err != nil {
-		t.Fatal(err)
-	}
-	if normalized.Status != GoalStatusRunning || normalized.StopCause != "" || normalized.RequestsUsed != 12 {
-		t.Fatalf("normalized sidecar = %+v", normalized)
-	}
-}
-
-func TestGoalSidecarPreservesUnknownFieldsAndOldReaderIgnoresRequests(t *testing.T) {
-	dir := t.TempDir()
-	path := filepath.Join(dir, "session.jsonl")
-	raw := []byte(`{"goal":"ship","status":"running","turnsLimit":10,"requestsUsed":7,"futurePolicy":{"mode":"adaptive"}}`)
-	if err := os.WriteFile(store.SessionGoalState(path), raw, 0o600); err != nil {
-		t.Fatal(err)
-	}
-	g := &goalMachine{}
-	g.restoreFromState(path)
-	g.mu.Lock()
-	_, data, ok := g.buildStateLocked(nil)
-	g.mu.Unlock()
-	if !ok {
-		t.Fatal("expected persisted state")
-	}
-	var roundTrip map[string]json.RawMessage
-	if err := json.Unmarshal(data, &roundTrip); err != nil {
-		t.Fatal(err)
-	}
-	if string(roundTrip["futurePolicy"]) != `{"mode":"adaptive"}` {
-		t.Fatalf("unknown field was lost: %s", data)
-	}
-	var oldReader struct {
-		Goal       string `json:"goal"`
-		TurnsLimit int    `json:"turnsLimit"`
-	}
-	if err := json.Unmarshal(data, &oldReader); err != nil || oldReader.Goal != "ship" || oldReader.TurnsLimit != 10 {
-		t.Fatalf("old reader = %+v err=%v", oldReader, err)
-	}
-}
-
-func TestGoalRunBoundariesPauseAfterTerminalDecisions(t *testing.T) {
-	t.Run("continue pauses", func(t *testing.T) {
-		g := &goalMachine{goal: "ship", status: GoalStatusRunning, turnsLimit: 10}
-		res := g.advance(goalAdvanceInput{
-			report:     &goalTurnReport{status: GoalStatusRunning, reason: "more remains"},
-			pauseCause: stopCauseGoalRunBudget, pauseReason: "16 model rounds used",
-		})
-		if res.cont || g.status != GoalStatusBlocked || g.stopCause != stopCauseGoalRunBudget {
-			t.Fatalf("result=%+v runtime=%+v", res, g.runtimeView())
-		}
-		before := g.turnsLimit
-		_, _, _, resumed, extended := g.resume(nil)
-		if !resumed || extended || g.turnsLimit != before {
-			t.Fatalf("resume = resumed:%v extended:%v turns:%d/%d", resumed, extended, g.turnsUsed, g.turnsLimit)
-		}
-	})
-
-	t.Run("completion wins", func(t *testing.T) {
-		g := &goalMachine{goal: "ship", status: GoalStatusRunning, turnsLimit: 10}
-		res := g.advance(goalAdvanceInput{
-			report:     &goalTurnReport{status: GoalStatusComplete},
-			readiness:  agent.ReadinessResult{Ready: true},
-			pauseCause: stopCauseGoalStuck, pauseReason: "repeated host failure",
-		})
-		if res.notice != goalCompleteNotice || g.status != GoalStatusComplete || g.stopCause != "" {
-			t.Fatalf("result=%+v runtime=%+v", res, g.runtimeView())
-		}
-	})
-}
-
-func TestGoalProgressEvidenceRestoresAsBoundedNoveltyState(t *testing.T) {
-	dir := t.TempDir()
-	sessionPath := filepath.Join(dir, "session.jsonl")
-	state := goalState{
-		Goal: "research the protocol", Status: GoalStatusRunning,
-		BudgetClass: budgetClassResearch, TurnsLimit: budgetQuota(budgetClassResearch),
-		NoProgressTurns: 2, NoProgressLimit: defaultNoProgressLimit,
-		ProgressEvidence: []string{"read-a", "read-a", "read-b", strings.Repeat("x", 129)},
-	}
-	raw, err := json.Marshal(state)
-	if err != nil {
-		t.Fatal(err)
-	}
-	if err := writeGoalStateData(goalStatePath(sessionPath), raw); err != nil {
-		t.Fatal(err)
-	}
-
-	g := &goalMachine{}
-	_, migratedData, migrated, _ := g.restoreFromState(sessionPath)
-	if !migrated {
-		t.Fatal("legacy flat research limit must migrate to the class quota")
-	}
-	if g.noProgressLimit != 10 || len(g.progressEvidence) != 2 {
-		t.Fatalf("restored runtime = limit:%d evidence:%v", g.noProgressLimit, g.progressEvidence)
-	}
-	var normalized goalState
-	if err := json.Unmarshal(migratedData, &normalized); err != nil {
-		t.Fatal(err)
-	}
-	if normalized.NoProgressLimit != 10 || len(normalized.ProgressEvidence) != 2 {
-		t.Fatalf("normalized sidecar = limit:%d evidence:%v", normalized.NoProgressLimit, normalized.ProgressEvidence)
-	}
-
-	// A restored repeat is still zero-gain; a genuinely new identity resets the
-	// stall counter without requiring any mutation.
-	g.advance(goalAdvanceInput{
-		report:           &goalTurnReport{status: GoalStatusRunning, reason: "repeat"},
-		progressEvidence: []string{"read-a"},
-	})
-	if g.noProgressTurns != 3 {
-		t.Fatalf("restored repeat reset no-progress to %d", g.noProgressTurns)
-	}
-	g.advance(goalAdvanceInput{
-		report:           &goalTurnReport{status: GoalStatusRunning, reason: "new result"},
-		progressEvidence: []string{"read-c"},
-	})
-	if g.noProgressTurns != 0 {
-		t.Fatalf("new restored evidence left no-progress at %d", g.noProgressTurns)
 	}
 }
