+++ internal/evidence/evidence.go
@@ -2,7 +2,6 @@ package evidence
 
 import (
 	"context"
-	"crypto/sha256"
 	"encoding/json"
 	"fmt"
 	"path/filepath"
@@ -524,68 +523,6 @@ func (l *Ledger) HasWriteOrCommandSince(index int) bool {
 		}
 	}
 	return false
-}
-
-// SuccessfulProgressSignaturesSince returns stable identities for successful
-// host-observed work recorded at or after index. Callers can keep a scoped set
-// of these signatures so new reads/results, task-state changes, reviews, and
-// mutations renew an execution lease while exact repeats do not masquerade as
-// progress.
-func (l *Ledger) SuccessfulProgressSignaturesSince(index int) []string {
-	if l == nil {
-		return nil
-	}
-	if index < 0 {
-		index = 0
-	}
-	l.mu.Lock()
-	defer l.mu.Unlock()
-	var out []string
-	for i := index; i < len(l.receipts); i++ {
-		if sig, ok := progressReceiptSignature(l.receipts[i]); ok {
-			out = append(out, sig)
-		}
-	}
-	return out
-}
-
-func progressReceiptSignature(r Receipt) (string, bool) {
-	if !r.Success {
-		return "", false
-	}
-	kind := ""
-	switch {
-	case r.Mutation || r.Write:
-		kind = "mutation"
-	case r.Command != "":
-		kind = "command"
-	case r.ToolName == "todo_write" && len(r.Todos) > 0:
-		kind = "todo"
-	case r.ToolName == "complete_step" && r.StepProof:
-		kind = "signoff"
-	case successfulForegroundReviewReceipt(r) || completedStructuredReviewReceipt(r, nil):
-		kind = "review"
-	case r.Read && r.OutputBytes > 0:
-		kind = "read"
-	default:
-		return "", false
-	}
-	payload := strings.TrimSpace(string(r.Args))
-	var decoded any
-	if json.Unmarshal(r.Args, &decoded) == nil {
-		if canonical, err := json.Marshal(decoded); err == nil {
-			payload = string(canonical)
-		}
-	}
-	// Result identity matters for observational work: the same query producing
-	// changed content is new evidence, while the same query and output is not.
-	// Mutations and state transitions remain argument/state keyed so an exact
-	// successful replay cannot keep a Goal alive by itself.
-	if (kind == "read" || kind == "command") && r.OutputDigest != "" {
-		payload += "\x00output=" + r.OutputDigest
-	}
-	sum := sha256.Sum256([]byte(kind + "\x00" + r.ToolName + "\x00" + payload))
-	return fmt.Sprintf("%x", sum), true
 }
 
 func (l *Ledger) HasSuccessfulCommand(command string) bool {
