+++ internal/provider/openai/lifecycle.go
@@ -1,0 +1,8 @@
+package openai
+
+// CloseIdleConnections releases pooled HTTP connections for bounded probes.
+func (c *client) CloseIdleConnections() {
+	if c != nil && c.http != nil {
+		c.http.CloseIdleConnections()
+	}
+}
